requirements.txt
numpy==1.17.1
Pillow==6.1.0
pip-autoremove==0.9.1
six==1.12.0
torch==1.2.0
torchvision==0.4.0
wheezy.captcha==0.1.44
